add inflation, reduce warnings

+++ path_searching/include/path_searching/path_node.h
@@ -6,8 +6,8 @@
 #include <iostream>
 #include <memory>
 #include <queue>
-#include <vector>
 #include <unordered_map>
+#include <vector>
 
 class PathNode;
 typedef PathNode* PathNodePtr;
@@ -37,7 +37,7 @@ template <typename T>
 struct MatrixHash : std::unary_function<T, size_t> {
   std::size_t operator()(T const& matrix) const {
     size_t seed = 0;
-    for (size_t i = 0; i < matrix.size(); ++i) {
+    for (size_t i = 0; i < (long unsigned int)(matrix.size()); ++i) {
       auto elem = *(matrix.data() + i);
       seed ^= std::hash<typename T::Scalar>()(elem) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
     }
@@ -73,6 +73,5 @@ class NodeHashTable {
     data_4d_.clear();
   }
 };
-
 
 #endif  // _PATH_NODE_H_
